pyproject.toml
[tool.poetry]
name = "kur-6"
version = "0.1.0"
description = ""
authors = ["Nikita <nikita@mer1d1an.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2 = "^2.9.9"
ipython = "^8.24.0"
pillow = "^10.3.0"
slugify = "^0.0.1"
django-apscheduler = "^0.6.2"
python-dotenv = "^1.0.1"
redis = "^5.0.7"
black = "^24.4.2"
pytils = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
